+++ packages/collections-domain/src/RecollectionDecorator.php
@@ -218,34 +218,18 @@ class RecollectionDecorator implements Recollection
     }
 
     /**
-     * @param null|Collection<TKey,T> $collection
-     * @param null|non-empty-array<string,Order>|string $orderBy
-     * @param null|int<1,max> $itemsPerPage
-     * @param null|int<0,max> $count
-     * @param null|int<1,max> $softLimit
-     * @param null|int<1,max> $hardLimit
+     * @param int<1,max> $itemsPerPage
      */
-    protected function with(
-        ?Collection $collection = null,
-        array|string|null $orderBy = null,
-        ?int $itemsPerPage = null,
-        ?CountStrategy $countStrategy = null,
-        ?int &$count = null,
-        ?int $softLimit = null,
-        ?int $hardLimit = null,
-    ): static {
-        $count = $count ?? $this->count;
-
-        // @phpstan-ignore-next-line
+    public function withItemsPerPage(int $itemsPerPage): static
+    {
         return static::create(
-            collection: $collection ?? $this->collection,
-            orderBy: $orderBy ?? $this->orderBy,
-            indexBy: $this->indexBy,
-            itemsPerPage: $itemsPerPage ?? $this->itemsPerPage,
-            countStrategy: $countStrategy ?? $this->countStrategy,
-            count: $count,
-            softLimit: $softLimit ?? $this->softLimit,
-            hardLimit: $hardLimit ?? $this->hardLimit,
+            collection: $this->collection,
+            orderBy: $this->orderBy,
+            itemsPerPage: $itemsPerPage,
+            countStrategy: $this->countStrategy,
+            count: $this->count,
+            softLimit: $this->softLimit,
+            hardLimit: $this->hardLimit,
         );
     }