+++ tests/src/IntegrationTests/RecollectionDecoratorSmallSetTest.php
@@ -30,6 +30,11 @@ class RecollectionDecoratorSmallSetTest extends BaseRecollectionTestCase
     /** @use RecollectionTestsTrait<Recollection<array-key,Citizen>> */
     use RecollectionTestsTrait;
 
+    protected function isSingleton(): bool
+    {
+        return true;
+    }
+
     protected function getExpectedTotal(): int
     {
         return 10;